Prepare command registry for plugin mechanism
Read more #13
